Setting up for an oncoming branch for the title screen.
Haven't got into the logistics of what will be the final spec, but the
files have been renamed for the work in progress move of the
implementation of the title screen code to work with viewports.

Also, set up a new test control for repeated firing of actions, also
called polling. This code is still very WIP in will be implemented into
a global file of utilities for handling all input processing, at a later
date.