Added the rest of the images to the test tileset, so I could draw the window.
Added the ability for the game to detect collisions with both floors and walls.

TODO: Still need to add a tileset implementation for out map, so we can use one image for all tiles, etc. Still need to implement player jump and still need to play around with final speed and gravity values.