Lots of changes here. Changed name to current working name. Also changed over sprites.
There is a lot of more changes to come but things that have been changed and worked out since last commit:

-Title Screen Refactoring
-Code Commenting
-Collision Detection
-GameWorld Orientation Changed (Top Down now, was 2d platformer before.
-Lots more

Git workflow should be as of working in Dev Branch, Make new branch for feature you are working on. Commit in there every few hours, merge branch and delete it as you finish that feature. We need to track our dev better so this is what programmers should strive for.