Cleaned up the sketchy import of the new map.
Implemented the map to load when we begin the game.

Added code to have the character be followed by the camera. This is only temporary as we will find a better way to handle this at a later date but for now will allow us to follow him throughout the whole map during testing. Who knows, we might end up liking it be the way it is.

TODO: Map loader for the game screen
TODO: make a way for the map loader to find any events that should take place on currently loaded map. Also have it load those events into the collision manager. We will solve this by programming into the map loader a way for it to find events and also have it put those points into an array. We will then be able to pass those events to the right method based off the name of the properties it finds.