Added animation logic. The animation isn't flawless, but it changes based on what direction the player is moving. We can change how fast the animation plays at any moment in time. This should close the Issue #8. Some other things have probably also been fixed.
I know for one fact a bug that never actually got thrown in the issue tracker had gotten fixed. I don't remember if I have committed since then but oh well. It's somewhere in the log now. You can move all directions regardless of if a wall is in front of you when you press the movement key.